+++ packages/apptive_grid_form/lib/widgets/apptive_grid_form_widgets.dart
@@ -19,6 +19,7 @@ part 'form_widget/form_widget.dart';
 part 'form_widget/attachment_form_widget.dart';
 part 'form_widget/check_box_form_widget.dart';
 part 'form_widget/enum_form_widget.dart';
+part 'form_widget/enum_collection_form_widget.dart';
 part 'form_widget/date_time_form_widget.dart';
 part 'form_widget/date_form_widget.dart';
 part 'form_widget/text_form_widget.dart';