+++ packages/apptive_grid_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.8.0-alpha.3]
+* **BREAKING CHANGE** `DataType.selectionBox` is now `DataType.singleSelect`
+* Add support for `DataType.enumCollection`
+
 ## [0.8.0-alpha.2]
 * Include Fixes from [0.7.1]